requirements.txt
Django==2.2.4
django-filter==2.2.0
djangorestframework==3.10.2
pytz==2019.2
sqlparse==0.3.0
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Django==2.2.4
django-filter==2.2.0
djangorestframework==3.10.2
firebase-admin==2.17.0
google-api-core==1.14.2
google-api-python-client==1.7.10
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.4.0
google-cloud-storage==1.18.0
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
grpcio==1.22.0
httplib2==0.13.1
idna==2.8
msgpack==0.6.1
protobuf==3.9.1
psycopg2-binary==2.8.3
pyasn1==0.4.6
pyasn1-modules==0.2.6
pytz==2019.2
requests==2.22.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
flake8==3.5.0
gunicorn
pyjwt
werkzeug
django-cors-headers